XMLHttpRequest/tests/submissions/Opera/send-entity-body-none.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send(null) - no entity body</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("POST", "resources/content.php", false)
        client.send(null)
        assert_equals(client.getResponseHeader("x-request-content-length"), "NO")
        assert_equals(client.getResponseHeader("x-request-content-type"), "NO")
      })
    </script>
  </body>
</html>
